Master staged

See merge request !91

+++ BudgetBunny/src/main/java/com/revature/controller/BudgetController.java
@@ -6,6 +6,7 @@ import java.util.ArrayList;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -23,6 +24,7 @@ import com.revature.service.CategoryService;
 public class BudgetController {
 	@Autowired
 	CategoryService categoryService;
+	Logger log = Logger.getRootLogger();
 	
 	@RequestMapping(method=RequestMethod.GET)
 	public String getBudgetPage(HttpServletRequest req)
@@ -42,9 +44,6 @@ public class BudgetController {
 	{
 		String category = req.getParameter("categoryData");
 		String deleted = req.getParameter("deletedList");
-		
-		System.out.println(req.getParameter("categoryData"));
-		System.out.println(deleted);
 		
 		ArrayList<Category> cList = new ArrayList<Category>();
 		ArrayList<Integer> dList = new ArrayList<Integer>();
@@ -71,12 +70,20 @@ public class BudgetController {
 				}			
 			}
 		} catch (JsonProcessingException e) {
-			e.printStackTrace();
+			log.error(e.getMessage());
 		} catch (IOException e) {
-			e.printStackTrace();
+			log.error(e.getMessage());
 		}
 		
-		categoryService.deleteList(dList);
+		System.out.println(dList);
+		if(dList.size() > 0)
+		{
+			double spent = categoryService.deleteList(dList);
+			user.getBudget().setTotalBudget(user.getBudget().getTotalBudget() + spent);
+			user.getBudget().setTotalSpent(user.getBudget().getTotalSpent() - spent);
+			req.getSession().setAttribute("user", user);
+		}
+		
 		categoryService.mergeList(cList);
 		
 		return "budgetpage";
